+++ CHANGELOG.md
@@ -74,6 +74,7 @@
 * Invoice field mapping: Exchange rate checkbox is not being saved. Fixed (UIDATIMP-864)
 * Invoice field mapping profile: Subscription start date field misaligned (UIDATIMP-850)
 * Match profile: form values are reset on page resize (UIDATIMP-867)
+* Fix Accessibility problems for settings/data-import/{...-profiles} (id attribute value must be unique) (UIDATIMP-462)
 
 ## [3.0.3](https://github.com/folio-org/ui-data-import/tree/v3.0.3) (2020-11-13)